Mostly bug fixes and some reorganization.
Fixed a bug that caused instances to be initialized multiple times. Fixed a bug that was causting instances to be
assigned to jQuery objects that already had them. The plugin used to attach to forms and form fields, it now strictly
attaches to forms.

Moved "private" functions out of the class and into the closure where they are now truly private.