Updated instance attachment and callbacks.
Instance attachment will now override previously stored instances. This is handy if you want to re-initialize
the plugin. For this reason, options merging is now handled within initialize, instead of before attachment.

Moved callback methods back into the autosave class. Two reasons for this: when they are public, they can be
accessed inside of other callback methods more intuitively, and also because in their previous position in the
code they didn't have access to some of the other private functions (getForms, getFields, etc). Callback
methods are now parsed and stored in their original position (options.save) instead of being stored in a new
object.

Reverted the rename of the "data" option to "extract".