Simplify and refactor.
Simplify the jQuery function. It now just creates an instance and passes off all the hard labor to the
autosave class.

Move the getter methods back into the class scope since they may be useful to someone and there's no sense in
hiding them.  They cannot be called statically anymore, though. Add getter methods 'getValidFields' and
'getChangedFields' for convenience.

Implement an 'update' function that will re-scan the original elements for changes before saving. This is done
in case anything is added to the DOM dynamically. This function may also be called by the user if needed.